chore: changing Notifications Component structure
